# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Grocery App",
        "description": "Grocery App",
        "license": {}
    },
    "host": "localhost:4000",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/api/product": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all Products",
                "parameters": [
                    {
                        "name": "productName",
                        "in": "query",
                        "required": false,
                        "description": "Search by Product Name",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ServiceResponse2"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create Product API",
                "parameters": [
                    {
                        "name": "productName",
                        "in": "formData",
                        "description": "Product Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "category",
                        "in": "formData",
                        "description": "Product Category",
                        "type": "string"
                    },
                    {
                        "name": "productShortDescription",
                        "in": "formData",
                        "description": "Product Short Description",
                        "type": "string"
                    },
                    {
                        "name": "productDescription",
                        "in": "formData",
                        "description": "Product Description",
                        "type": "string"
                    },
                    {
                        "name": "productPrice",
                        "in": "formData",
                        "description": "Product Price",
                        "type": "string"
                    },
                    {
                        "name": "productSalePrice",
                        "in": "formData",
                        "description": "Product Sale Price",
                        "type": "string"
                    },
                    {
                        "name": "productSKU",
                        "in": "formData",
                        "description": "Product SKU",
                        "type": "string"
                    },
                    {
                        "name": "productType",
                        "in": "formData",
                        "description": "Product Type",
                        "type": "string"
                    },
                    {
                        "name": "stockStatus",
                        "in": "formData",
                        "description": "Product Stock Status",
                        "type": "string"
                    },
                    {
                        "name": "productImage",
                        "in": "formData",
                        "description": "Product Image",
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ServiceResponse2"
                        }
                    }
                }
            }
        },
        "/api/product/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "Product Id",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Product Details based on Product Id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ServiceResponse2"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update Product API",
                "parameters": [
                    {
                        "name": "productName",
                        "in": "formData",
                        "description": "Product Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "category",
                        "in": "formData",
                        "description": "Product Category",
                        "type": "string"
                    },
                    {
                        "name": "productShortDescription",
                        "in": "formData",
                        "description": "Product Short Description",
                        "type": "string"
                    },
                    {
                        "name": "productDescription",
                        "in": "formData",
                        "description": "Product Description",
                        "type": "string"
                    },
                    {
                        "name": "productPrice",
                        "in": "formData",
                        "description": "Product Price",
                        "type": "string"
                    },
                    {
                        "name": "productSalePrice",
                        "in": "formData",
                        "description": "Product Sale Price",
                        "type": "string"
                    },
                    {
                        "name": "productSKU",
                        "in": "formData",
                        "description": "Product SKU",
                        "type": "string"
                    },
                    {
                        "name": "productType",
                        "in": "formData",
                        "description": "Product Type",
                        "type": "string"
                    },
                    {
                        "name": "stockStatus",
                        "in": "formData",
                        "description": "Product Stock Status",
                        "type": "string"
                    },
                    {
                        "name": "productImage",
                        "in": "formData",
                        "description": "Product Image",
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ServiceResponse2"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete Product API",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ServiceResponse2"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Products": {
            "required": [
                "_id",
                "productName",
                "category",
                "productPrice",
                "productType"
            ],
            "properties": {
                "_id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "productName": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "productShortDescription": {
                    "type": "string"
                },
                "productDescription": {
                    "type": "string"
                },
                "productPrice": {
                    "type": "string"
                },
                "productSalePrice": {
                    "type": "string"
                },
                "productSKU": {
                    "type": "string"
                },
                "productType": {
                    "type": "string"
                },
                "stockStatus": {
                    "type": "string"
                },
                "productImage": {
                    "type": "string"
                }
            }
        },
...
        "ServiceResponse2": {
            "properties": {
                "message": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Products"
                    }
                }
            }
        }
    }
}